upd DWO case
