+++ Validation/DWO/exp_data/050O.csv
@@ -1,0 +1,67 @@
+x, y
+307.1428843906948, -6060629413513603
+322.32142857142856, 0
+315.17861230032787, -3030256908311076.5
+329.4643129621233, 3030372505202311
+340.1785714285714, 18181888240540164
+340.1785714285714, 18181888240540164
+349.99998637608115, 54545433527838020
+346.42861230032787, 42424290297702480
+359.8214694431849, 90909094412026900
+362.5, 93939466917229420
+362.5, 93939466917229420
+369.64288439069475, 136363641618040450
+375.0000136239188, 160606043675202750
+380.35714285714283, 203030333972905250
+384.82142857142856, 227272736030067550
+391.071469443185, 233333365443580930
+394.6428435189383, 281818169557905570
+404.46432658604215, 309090944120269950
+402.67857142857144, 312121201028581060
+414.2857415335519, 269697026327770020
+418.75002724783764, 224242479121756480
+424.9999999999999, 184848561329256320
+429.4642857142857, 151515157353378300
+434.82141494750977, 124242498387904910
+444.64289801461354, 112121239560878130
+451.78571428571433, 93939466917229420
+457.1428435189383, 78787951181891570
+463.3928843906947, 78787951181891570
+474.9999863760812, 69696949263175870
+483.0357142857142, 54545433527838020
+495.535727909633, 45454547206013550
+505.3571428571428, 45454547206013550
+511.6071837288992, 33333403975878012
+518.75, 42424290297702480
+526.785727909633, 27272774562364624
+534.8214558192662, 33333403975878012
+546.4285578046525, 24242402057162316
+553.5714421953473, 33333403975878012
+568.7499863760811, 12121258827026774
+568.7499863760811, 12121258827026774
+577.6785578046527, 12121258827026774
+588.3928843906947, 6060629413513387
+598.2142993382045, 3030372505202311
+605.3571837288991, 3030372505202311
+611.6071564810616, 18181888240540164
+619.6428843906947, 9090886321824464
+619.6428843906947, 9090886321824464
+626.7857006617953, -3030256908311076.5
+634.8214285714286, 6060629413513387
+647.3214421953473, 0
+647.3214421953473, 0
+656.2500136239188, -9090886321824678
+663.3928980146135, 9090886321824464
+672.3214694431849, -3030256908311076.5
+688.3928571428571, -3030256908311076.5
+696.4285850524902, 12121258827026774
+704.4643129621232, 6060629413513387
+713.3928843906947, 0
+721.4286123003278, 0
+727.6785850524901, 0
+744.6428843906947, 3030372505202311
+744.6428843906947, 3030372505202311
+755.3571428571429, -3030256908311076.5
+764.2857142857142, 3030372505202311
+781.2500136239188, 12121258827026774
+791.9642720903668, 0
