+++ Validation/DWO/exp_data/075O.csv
@@ -1,0 +1,67 @@
+x, y
+310.7143265860421, 3030256324497606
+325.89287076677596, 6060743842733139
+325.89287076677596, 6060743842733139
+337.5000408717564, 3030256324497606
+336.60712923322404, 21212256658957804
+341.07141494750977, 24242512983454980
+346.42861230032787, 21212256658957804
+351.78574153355186, 24242512983454980
+354.46427209036693, 27272769307952584
+354.46427209036693, 27272769307952584
+361.6071564810616, 51515282291407570
+361.6071564810616, 51515282291407570
+369.64288439069475, 45454538448674856
+379.4642993382045, 78787820405621800
+384.82142857142856, 72727307756627000
+384.82142857142856, 72727307756627000
+391.071469443185, 72727307756627000
+398.2142857142857, 63636307589396696
+405.3571701049804, 51515282291407570
+416.07142857142856, 39394025799680070
+416.07142857142856, 39394025799680070
+429.4642857142857, 48484794773172030
+429.4642857142857, 48484794773172030
+434.82141494750977, 33333281956946936
+443.7499863760811, 18181769140722270
+443.7499863760811, 18181769140722270
+456.25, 18181769140722270
+460.71428571428567, 39394025799680070
+470.53570066179554, 21212256658957804
+478.57142857142856, 33333281956946936
+484.8214694431849, 18181769140722270
+495.535727909633, 21212256658957804
+503.57145581926613, 12121256491727920
+516.0714694431849, 15151512816225096
+516.0714694431849, 15151512816225096
+525.0000408717564, 18181769140722270
+535.7142993382045, 12121256491727920
+545.5357142857142, 24242512983454980
+553.5714421953473, 9091000167230314
+562.5000136239188, 12121256491727920
+571.4285850524902, 9091000167230314
+585.7142857142857, 9091000167230314
+593.7500136239187, 15151512816225096
+600.8928980146136, 9091000167230314
+609.8214694431849, 0
+618.7500408717565, 9091000167230314
+625.0000136239187, 9091000167230314
+632.1428980146135, 3030256324497606
+638.3928707667759, 0
+648.2142857142856, 0
+657.1428571428571, 0
+669.6428707667759, 3030256324497606
+669.6428707667759, 3030256324497606
+677.6785986764089, 3030256324497606
+694.6428980146134, -15151512816224666
+698.2142720903669, 9091000167230314
+705.3571564810616, -15151512816224666
+714.2857279096329, 3030256324497606
+721.4286123003278, -15151512816224666
+729.4642720903669, 0
+744.6428843906947, -3030256324497175.5
+754.4642993382045, 3030256324497606
+761.6071837288991, -3030256324497175.5
+767.8571564810616, 6060743842733139
+776.7857279096331, 3030256324497606
+783.0357006617953, -6060512648994781
